requirements.txt
altgraph==0.17
asgiref==3.2.10
asn1crypto==0.24.0
beautifulsoup4==4.9.1
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==7.1.2
crcmod==1.7
cryptography==2.6.1
defusedxml==0.7.0rc1
Django==3.0.8
django-extra-fields==2.0.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
entrypoints==0.3
fasteners==0.12.0
future==0.16.0
httplib2==0.11.3
idna==2.6
keyring==18.0.1
keyrings.alt==3.1.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
Mako==1.0.7
MarkupSafe==1.1.0
monotonic==1.5
MouseInfo==0.1.3
mysqlclient==2.0.1
netifaces==0.10.4
oauthlib==3.0.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
protobuf==3.12.0
pycairo==1.16.2
pycrypto==2.6.1
Pygments==2.6.1
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyperclip==1.8.0
pyRFC3339==1.1
PyScreeze==0.1.26
python-dateutil==2.7.3
python-debian==0.1.36
python3-openid==3.2.0
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.2
pyxdg==0.25
PyYAML==5.1.2
requests==2.21.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
stringify==0.1.1
urllib3==1.24.1
wadllib==1.3.3
django-storages[google]==1.9.1
google-cloud-storage==1.30.0